+++ src/main/java/domain/card/Deck.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import java.util.ArrayDeque;
 import java.util.Deque;
