+++ src/main/java/domain/player/HandsState.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.player;
 
 import java.util.Arrays;
 import java.util.function.Predicate;
