+++ src/main/java/view/OutputView.java
@@ -1,7 +1,7 @@
 package view;
 
-import domain.Card;
-import domain.Result;
+import domain.card.Card;
+import domain.result.Result;
 
 import java.util.List;
