+++ src/test/java/domain/DeckTest.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.card.Deck;
+import domain.card.RandomCardGenerator;
 import org.assertj.core.api.InstanceOfAssertFactories;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
