+++ src/test/java/domain/GameResultTest.java
@@ -1,5 +1,12 @@
 package domain;
 
+import domain.card.Card;
+import domain.card.Deck;
+import domain.card.Number;
+import domain.card.Shape;
+import domain.player.Dealer;
+import domain.player.Participants;
+import domain.result.Result;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
